tr1/game-flow: drop level_id from level-stats
