+++ data/tr1/ship/cfg/TR1X_gameflow.json5
@@ -63,7 +63,7 @@
                 {"type": "loading_screen", "path": "data/images/gym.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 0},
+                {"type": "level_stats"},
                 {"type": "exit_to_title"},
             ],
         },
@@ -82,7 +82,7 @@
                 {"type": "loading_screen", "path": "data/images/peru.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 1},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 2},
             ],
         },
@@ -99,7 +99,7 @@
                 {"type": "loading_screen", "path": "data/images/peru.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 2},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 3},
             ],
         },
@@ -118,7 +118,7 @@
                 {"type": "loading_screen", "path": "data/images/peru.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 3},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 4},
             ],
         },
@@ -138,7 +138,7 @@
                 {"type": "load_level"},
                 {"type": "play_level"},
                 {"type": "play_cutscene", "cutscene_id": 0},
-                {"type": "level_stats", "level_id": 4},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 5},
             ],
         },
@@ -157,7 +157,7 @@
                 {"type": "loading_screen", "path": "data/images/greece.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 5},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 6},
             ],
         },
@@ -177,7 +177,7 @@
                 {"type": "loading_screen", "path": "data/images/greece.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 6},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 7},
             ],
         },
@@ -194,7 +194,7 @@
                 {"type": "loading_screen", "path": "data/images/greece.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 7},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 8},
             ],
             "unobtainable_pickups": 1,
@@ -214,7 +214,7 @@
                 {"type": "loading_screen", "path": "data/images/greece.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 8},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 9},
             ],
         },
@@ -237,7 +237,7 @@
                 {"type": "load_level"},
                 {"type": "play_level"},
                 {"type": "play_cutscene", "cutscene_id": 1},
-                {"type": "level_stats", "level_id": 9},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 10},
             ],
         },
@@ -256,7 +256,7 @@
                 {"type": "loading_screen", "path": "data/images/egypt.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 10},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 11},
             ],
         },
@@ -276,7 +276,7 @@
                 {"type": "loading_screen", "path": "data/images/egypt.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 11},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 12},
             ],
         },
@@ -294,7 +294,7 @@
                 {"type": "loading_screen", "path": "data/images/egypt.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 12},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 13},
             ],
         },
@@ -326,7 +326,7 @@
                 {"type": "remove_scions"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 13},
+                {"type": "level_stats"},
                 {"type": "play_cutscene", "cutscene_id": 2},
                 {"type": "exit_to_level", "level_id": 14},
             ],
@@ -349,7 +349,7 @@
                 {"type": "play_level"},
                 {"type": "play_fmv", "fmv_id": 9},
                 {"type": "play_cutscene", "cutscene_id": 3},
-                {"type": "level_stats", "level_id": 14},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 15},
             ],
             "unobtainable_pickups": 3,
@@ -368,7 +368,7 @@
                 {"type": "loading_screen", "path": "data/images/atlantis.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 15},
+                {"type": "level_stats"},
                 {"type": "play_fmv", "fmv_id": 10},
                 {"type": "play_music", "music_track": 19},
                 {"type": "display_picture", "path": "data/images/end.webp", "display_time": 7.5},
@@ -417,7 +417,7 @@
                 {"type": "loading_screen", "path": "data/images/peru.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 2},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 3},
             ],
         },
@@ -436,7 +436,7 @@
                 {"type": "loading_screen", "path": "data/images/peru.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 3},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 4},
             ],
         },
