+++ data/tr1/ship/cfg/TR1X_gameflow_demo_pc.json5
@@ -50,7 +50,7 @@
                 {"type": "loading_screen", "path": "data/images/peru.webp"},
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 0},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 1},
             ],
             "demo": true,
