+++ data/tr1/ship/cfg/TR1X_gameflow_ub.json5
@@ -57,7 +57,7 @@
             "sequence": [
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 0},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 1},
             ],
             "strings": {"key1": "Gold Key"},
@@ -78,7 +78,7 @@
             "sequence": [
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 1},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 2},
             ],
             "strings": {"key1": "Ornate Key"},
@@ -97,7 +97,7 @@
             "sequence": [
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 2},
+                {"type": "level_stats"},
                 {"type": "exit_to_level", "level_id": 3},
             ],
             "unobtainable_kills": 1,
@@ -115,7 +115,7 @@
             "sequence": [
                 {"type": "load_level"},
                 {"type": "play_level"},
-                {"type": "level_stats", "level_id": 3},
+                {"type": "level_stats"},
                 {"type": "play_music", "music_track": 19},
                 {"type": "display_picture", "path": "data/images/end.webp", "display_time": 7.5},
                 {"type": "display_picture", "path": "data/images/credits_ub.webp", "display_time": 7.5},
