+++ src/tr2/game/game_flow/sequencer.c
@@ -171,14 +171,11 @@ GAME_FLOW_COMMAND GF_InterpretSequence(
         }
 
         case GFS_LEVEL_STATS: {
-            if (type != GFL_NORMAL) {
-                break;
-            }
             const GAME_FLOW_LEVEL *const current_level = Game_GetCurrentLevel();
             PHASE *const stats_phase = Phase_Stats_Create((PHASE_STATS_ARGS) {
                 .background_type = BK_OBJECT,
-                .show_final_stats = false,
                 .level_num = current_level->num,
+                .show_final_stats = false,
                 .use_bare_style = false,
             });
             gf_cmd = PhaseExecutor_Run(stats_phase);
@@ -194,7 +191,7 @@ GAME_FLOW_COMMAND GF_InterpretSequence(
             START_INFO *const start = GF_GetResumeInfo(current_level);
             start->stats = g_SaveGame.current_stats;
             start->available = 0;
-            GAME_FLOW_LEVEL *const next_level =
+            const GAME_FLOW_LEVEL *const next_level =
                 GF_GetLevel(current_level->num + 1, current_level->type);
             if (next_level != NULL) {
                 CreateStartInfo(next_level);
